Update several fields from integer to float

+++ db/migrate/20200805013546_create_foods.rb
@@ -4,27 +4,27 @@ class CreateFoods < ActiveRecord::Migration[6.0]
       t.string :brand_name
       t.string :description
       t.string :serving_size
-      t.integer :servings_per_container
+      t.float :servings_per_container
       t.integer :calories
-      t.integer :total_fat
-      t.integer :saturated_fat, default: 0
-      t.integer :polyunsaturated_fat, default: 0
-      t.integer :monounsaturated_fat, default: 0
-      t.integer :trans_fat, default: 0
-      t.integer :cholesterol, default: 0
-      t.integer :sodium, default: 0
-      t.integer :total_carbohydrate
-      t.integer :dietary_fiber, default: 0
-      t.integer :total_sugars, default: 0
-      t.integer :added_sugars, default: 0
-      t.integer :sugar_alcohols, default: 0
-      t.integer :protein
-      t.integer :vitamin_a, default: 0
-      t.integer :vitamin_c, default: 0
-      t.integer :vitamin_d, default: 0
-      t.integer :calcium, default: 0
-      t.integer :iron, default: 0
-      t.integer :potassium, default: 0
+      t.float :total_fat
+      t.float :saturated_fat, default: 0
+      t.float :polyunsaturated_fat, default: 0
+      t.float :monounsaturated_fat, default: 0
+      t.float :trans_fat, default: 0
+      t.float :cholesterol, default: 0
+      t.float :sodium, default: 0
+      t.float :total_carbohydrate
+      t.float :dietary_fiber, default: 0
+      t.float :total_sugars, default: 0
+      t.float :added_sugars, default: 0
+      t.float :sugar_alcohols, default: 0
+      t.float :protein
+      t.float :vitamin_a, default: 0
+      t.float :vitamin_c, default: 0
+      t.float :vitamin_d, default: 0
+      t.float :calcium, default: 0
+      t.float :iron, default: 0
+      t.float :potassium, default: 0
 
       t.timestamps
     end
